bamv1 working with mpileup.gz
