+++ names.filelist
@@ -1,0 +1,2 @@
+outFile.DP10.mpileup.gz
+outFile.DP20.mpileup.gz
