+++ ploidy.file
@@ -1,0 +1,3 @@
+2x20
+2x10,4x10
+3x20
